Preparing to rewrite the code, seems to be in a state where it still works
